Use type-safe kinetics config and deploy

+++ parser/src/worker.rs
@@ -1,13 +1,14 @@
 use crate::environment::{parse_environment, Environment};
+use serde::{Deserialize, Serialize};
 use syn::{
     parse::{Parse, ParseStream},
     token, Ident, LitBool, LitInt, LitStr,
 };
 
-#[derive(Debug, Clone)]
+#[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct Worker {
     pub name: Option<String>,
-    pub concurrency: i16,
+    pub concurrency: u32,
     pub fifo: bool,
     pub environment: Environment,
 }
@@ -39,7 +40,7 @@ impl Parse for Worker {
                     worker.environment = parse_environment(input)?;
                 }
                 "concurrency" => {
-                    worker.concurrency = input.parse::<LitInt>()?.base10_parse::<i16>()?;
+                    worker.concurrency = input.parse::<LitInt>()?.base10_parse::<u32>()?;
                 }
                 "fifo" => {
                     worker.fifo = match input.parse::<LitBool>() {
